Added asIObjectType interface
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@238 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_generic.cpp
@@ -80,7 +80,7 @@ int asCGeneric::GetObjectTypeId()
 
 int asCGeneric::GetArgCount()
 {
-	return sysFunction->parameterTypes.GetLength();
+	return (int)sysFunction->parameterTypes.GetLength();
 }
 
 asBYTE asCGeneric::GetArgByte(asUINT arg)
